+++ tests/telemetry/test_metrics.py
@@ -73,13 +73,6 @@ def test_inc_counter(metrics_collector, meter_provider):
     assert all_metrics["counters"]["http_requests_total"]["value"] == 2
 
 
-def test_statusline_bake_counters_registered(metrics_collector):
-    all_metrics = metrics_collector.get_all_metrics()
-
-    assert all_metrics["counters"]["statusline_posts_succeeded_total"]["value"] == 0
-    assert all_metrics["counters"]["statusline_usage_gap_warnings_total"]["value"] == 0
-
-
 def test_autonomous_stuck_lifecycle_counter_registered(metrics_collector):
     metrics_collector.inc_counter("agent_lifecycle_autonomous_stuck_detected_total")